chore: Remove typescript-eslint packages to get them from Coralloy preset
Either the pnpm upgrade, some dep upgrades, or something else made the
error about the resolve problem with typescript-eslint go away. So,
we can remove the packages from the project and get them from Coralloy
preset, as long as we are still using the eslint patch. So, I also added
a clarifying comment about that.